% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staggeredGMM-package.R
\docType{package}
\name{staggeredGMM-package}
\alias{staggeredGMM}
\alias{staggeredGMM-package}
\title{staggeredGMM: GMM Estimation of Treatment Effects under Staggered Adoption}
\description{
Implements the GMM estimator of Arora and Bijani, "Estimating Treatment
Effects under Staggered Timing and Non-Spherical Errors", for
cohort-and-time-specific average treatment effects (CATTs) under staggered
treatment adoption. Three weighting variants are provided:
\code{\link[=gmm_staggered_I]{gmm_staggered_I()}} (pooled Toeplitz / "Eff"), \code{\link[=gmm_staggered_II]{gmm_staggered_II()}}
(cohort-specific Toeplitz / "HetVar"), and \code{\link[=gmm_staggered_III]{gmm_staggered_III()}} (fully
unrestricted within-cohort covariance / "FullCov"). All three support
optional baseline-covariate adjustment via \code{covar}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RishabhBijani/staggeredGMM}
  \item Report bugs at \url{https://github.com/RishabhBijani/staggeredGMM/issues}
}

}
\author{
\strong{Maintainer}: Rishabh Bijani \email{rishabhbijani@gmail.com}

Authors:
\itemize{
  \item Rishabh Bijani \email{rishabhbijani@gmail.com}
  \item Parush Arora \email{parush.arora@ashoka.edu.in}
}

}
\keyword{internal}
